backend/requirements.txt
Django>=5.0,<7.0
djangorestframework>=3.14
Pillow>=10.0
gunicorn>=21.0
